+++ src/cgame/cg_weapons.cpp
@@ -738,9 +738,7 @@ static void CG_DynamiteTrail(centity_t *ent, const weaponInfo_t *wi) {
     mult = 0.004f * (cg.time - ent->currentState.effect1Time) / 30000.0f;
     trap_R_AddLightToScene(
         origin, 320,
-        static_cast<float>(std::fabs(std::sin(
-            static_cast<double>(cg.time - ent->currentState.effect1Time) *
-            mult))),
+        std::fabs(std::sin((cg.time - ent->currentState.effect1Time) * mult)),
         1.0, 0.0, 0.0, 0, REF_FORCE_DLIGHT);
   } else {
     mult = 1 - ((cg.time - ent->trailTime) / 15500.0f);
